git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rebmix@91 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/rebmix/src/rngmixf.cpp
@@ -564,7 +564,7 @@ S0: while (fgets(line, 2048, fp) != NULL) {
 
             Error = isI < 1; if (Error) goto E0;
 
-            MixTheta = new CompnentDistribution* [c_ + 1];
+            MixTheta = new CompnentDistribution* [(unsigned int)(c_ + 1)];
 
             Error = NULL == MixTheta; if (Error) goto E0;
